requirements.txt
django
djangorestframework
django-cors-headers
django-environ
djangorestframework-simplejwt
stripe
whitenoise
gunicorn
uvicorn
dj-database-url
psycopg2-binary
drf_yasg